backend/requirements.txt
flask
flask-session
flask-cors
python-dotenv
mysql-connector-python
bcrypt
